refactor: reorganize context and provider folder structure
