+++ apps/web/src/contexts/auth/AuthProvider.tsx
@@ -1,0 +1,12 @@
+import { api } from "../../utils/api";
+import { AuthContext } from "./AuthContext";
+
+export function AuthProvider({ children }: { children: React.ReactNode }) {
+  const { data: user, isLoading } = api.auth.validate.useQuery();
+
+  return (
+    <AuthContext.Provider value={{ user, isLoading }}>
+      {children}
+    </AuthContext.Provider>
+  );
+}
