+++ apps/web/src/layouts/AuthLayout.tsx
@@ -1,5 +1,5 @@
 import { useEffect } from "react";
-import { useAuth } from "../providers/AuthProvider";
+import { useAuth } from "../contexts/auth/useAuth";
 import { useNavigate } from "react-router";
 import { Outlet } from "react-router";
 export function AuthLayout() {
